+++ archive/2026-05-cleanup/src/menipy/common/zold_detection.py
@@ -1,0 +1,1 @@
+"""Deprecated/Empty file."""